refactor: Product 페이지 코드 분리
